+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: (zkSync) Fix max spend with new `useBalanceAsMaxSpendable` hack
+
 ## 4.17.1 (2024-07-30)
 
 - fixed: (ETH) Fix incorrect protocol in blockbook-ws connection URLs